\name{exRasch}
\alias{exRasch}
\docType{data}
\title{ Example 1-class data }
\description{
  A small example data set for single-class (standard) Rasch model analysis.
}
\usage{data(exRasch)}
\format{
  A data frame with 100 observations on the following 10 variables.
  \describe{
    \item{\code{i1}}{a numeric vector}
    \item{\code{i2}}{a numeric vector}
    \item{\code{i3}}{a numeric vector}
    \item{\code{i4}}{a numeric vector}
    \item{\code{i5}}{a numeric vector}
    \item{\code{i6}}{a numeric vector}
    \item{\code{i7}}{a numeric vector}
    \item{\code{i8}}{a numeric vector}
    \item{\code{i9}}{a numeric vector}
    \item{\code{i10}}{a numeric vector}
  }
}
\source{
  Simulated to follow a 1-class (standard) Rasch model.
}
\seealso{ mixRasch }
\examples{
data(exRasch)
}
\keyword{datasets}
