\name{mixRasch-package}
\alias{mixRasch-package}
\docType{package}
\title{
Contains a function for estimating and evaluating mixture Rasch models using JMLE
}
\description{
The included function will estimate a mixture Rasch model using joint maximum likelihood estimation (JMLE). The estimation is based on a mixture partial credit model. Step parameters can be constrained to estimate a mixture rating scale model. Estimating a model with only one latent class accomplishes a standard Rasch analysis with JMLE.
}
\details{
\tabular{ll}{
Package: \tab mixRasch\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-03-13\cr
License: \tab GPL version 2 or newer\cr
}
This is an early version of the package. It currently contains only the mixRasch function. Please contact the author if you encounter any bugs or if you have questions or suggestions.
}
\author{
John T. Willse <jtwillse@uncg.edu>
}
\references{
Willse, J. T. (2009). Mixture Rasch models with joint maximum likelihood estimation. Presented at the National Council on Measurement in Education annual meeting. San Diego.
}
\keyword{ package }

